requirements.txt
astroid==1.4.9
beautifulsoup4==4.5.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.2
isort==4.2.5
jieba3k==0.35.1
lazy-object-proxy==1.2.2
lxml==3.7.2
mccabe==0.5.3
munch==2.0.4
newspaper3k==0.1.9
nltk==3.2.2
olefile==0.44
Pillow==4.0.0
pylint==1.6.4
python-dateutil==2.6.0
PyYAML==3.12
requests==2.12.5
requests-file==1.4.1
rethinkdb==2.3.0.post6
six==1.10.0
tldextract==2.0.2
wrapt==1.10.8
